requirements.txt
flask
pdfplumber
scikit-learn
nltk
pandas
google-generativeai
python-dotenv
flask-sqlalchemy
fpdf2
